+++ app/imports/ui/pages/JobSearchResult.jsx
@@ -8,69 +8,7 @@ import JobCard from '../components/JobCard';
 /** Renders the Page for job search results. */
 class JobSearchResult extends React.Component {
 
-  jobs = [
-    {
-      _id: 1,
-      title: 'Job 1',
-      description: 'Leverage agile frameworks to provide a robust synopsis for high level overviews.',
-      location: 'Campus Center',
-      pay: 12.44,
-      postDate: subDays(new Date(), 3),
-      skills: [
-        { _id: 1, name: 'technology' },
-        { _id: 2, name: 'IT' },
-        { _id: 3, name: 'programming' },
-      ],
-    },
-    {
-      _id: 2,
-      title: 'Job 11',
-      description: 'Bring to the table win-win survival strategies to ensure proactive domination.',
-      location: 'Food Court',
-      pay: 11.22,
-      postDate: subDays(new Date(), 1),
-      skills: [
-        { _id: 1, name: 'MS Word' },
-        { _id: 2, name: 'IT' },
-        { _id: 3, name: 'Typing' },
-      ],
-    },
-    {
-      _id: 3,
-      title: 'Job 123',
-      description: 'Capitalise on low hanging fruit to identify a ballpark value added activity to beta test.',
-      location: 'Library',
-      pay: 10.11,
-      postDate: subDays(new Date(), 10),
-      skills: [
-        { _id: 1, name: 'AWS' },
-        { _id: 2, name: 'Unix' },
-      ],
-    },
-    {
-      _id: 4,
-      title: 'Job 12',
-      description: 'Capitalise on low hanging fruit to identify a ballpark value added activity to beta test.',
-      location: 'Landscaping',
-      pay: 12.22,
-      postDate: subDays(new Date(), 3),
-      skills: [
-        { _id: 2, name: 'React' },
-        { _id: 3, name: 'UI Design' },
-      ],
-    },
-    {
-      _id: 5,
-      title: 'Job 135',
-      description: 'Capitalise on low hanging fruit to identify a ballpark value added activity to beta test.',
-      location: 'Athletic Center',
-      pay: 10.77,
-      postDate: subDays(new Date(), 22),
-      skills: [
-        { _id: 3, name: 'Project Management' },
-      ],
-    },
-  ];
+
   /** Bind 'this' so that a ref to the Form can be saved in formRef and communicated between render() and submit(). */
   constructor(props) {
     super(props);
